% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_remesas.R
\name{calcular_gini}
\alias{calcular_gini}
\title{Calcular el Coeficiente de Gini}
\usage{
calcular_gini(x)
}
\arguments{
\item{x}{Un vector numérico con los datos (por ejemplo, ingresos o montos de transacciones).}
}
\value{
Un valor numérico entre 0 y 1 que representa el coeficiente de Gini.
}
\description{
Esta función calcula el coeficiente de Gini para medir la desigualdad o
concentración de una distribución (muy útil para modelos de riesgo o análisis de remesas).
}
\examples{
ingresos <- c(100, 200, 300, 400, 1000)
calcular_gini(ingresos)

}
